pyproject.toml
[project]
name = "mit-15-376-homework-4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "kokoro>=0.9.4",
    "numpy>=2.2.6",
    "openai>=2.2.0",
    "openai-whisper>=20250625",
    "opencv-python>=4.12.0.88",
    "scipy>=1.16.2",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.1",
]
